Move orderToStep function, fix tests.
